Adicionar validacao remover produto

+++ lib/view/produtos/listar_produtos_page.dart
@@ -1,5 +1,7 @@
 import 'package:desenvolvimentomobile/helper/error.dart';
+import 'package:desenvolvimentomobile/model/pedido.dart';
 import 'package:desenvolvimentomobile/model/produto.dart';
+import 'package:desenvolvimentomobile/repositories/pedido_repository.dart';
 import 'package:desenvolvimentomobile/repositories/produto_repository.dart';
 import 'package:desenvolvimentomobile/routes/routes.dart';
 import 'package:desenvolvimentomobile/widgets/drawer.dart';
@@ -46,13 +48,23 @@ class _ListarProdutosPageState extends State<ListarProdutosPage> {
   }
 
   void _removerProduto(Produto produto) async {
-      try {
-      ProdutoRepository repository = ProdutoRepository();
-      await repository.remover(produto.id!).then((value) {
-        _refreshList();
-        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text('Produto ${produto.id} removido com sucesso.')));
-      });
+    try {
+      PedidoRepository pedidoRepository = PedidoRepository();
+      List<Pedido> pedidos =
+          await pedidoRepository.listarPorIdProduto(produto.id!);
+
+      if (pedidos.isNotEmpty) {
+        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+            content: Text(
+                'Não é possível excluir esse cliente pois ele tem pedidos.')));
+      } else {
+        ProdutoRepository repository = ProdutoRepository();
+        await repository.remover(produto.id!).then((value) {
+          _refreshList();
+          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+              content: Text('Produto ${produto.id} removido com sucesso.')));
+        });
+      }
     } catch (exception) {
       showError(context, "Erro removendo produto", exception.toString());
     }
@@ -87,39 +99,41 @@ class _ListarProdutosPageState extends State<ListarProdutosPage> {
 
   void _editItem(BuildContext context, int index) {
     Produto b = _lista[index];
-    Navigator.pushNamed(
-      context,
-      Routes.produtoEdit,
-      arguments: <String, int>{"id": b.id!}
-      ).then((value) => _refreshList());
+    Navigator.pushNamed(context, Routes.produtoEdit,
+        arguments: <String, int>{"id": b.id!}).then((value) => _refreshList());
   }
 
   void _removeItem(BuildContext context, int index) {
-    Produto b = _lista[index]; 
-    showDialog( context: context,
-      builder: (BuildContext context) => 
-          AlertDialog( 
-            title: const Text("Remover Produto"),
-            content: Text("Gostaria realmente de remover ${b.descricao}?"),
-            actions: [  TextButton(
-                              child: const Text("Não"), 
-                              onPressed: () { Navigator.of(context).pop(); }, 
-                            ),
-                        TextButton( child: const Text("Sim"), onPressed: () {
-                            _removerProduto(b);
-                            Navigator.of(context).pop();
-                          }, 
-                        ), 
-                      ],
-          )
-    );
+    Produto b = _lista[index];
+    showDialog(
+        context: context,
+        builder: (BuildContext context) => AlertDialog(
+              title: const Text("Remover Produto"),
+              content: Text("Gostaria realmente de remover ${b.descricao}?"),
+              actions: [
+                TextButton(
+                  child: const Text("Não"),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  },
+                ),
+                TextButton(
+                  child: const Text("Sim"),
+                  onPressed: () {
+                    _removerProduto(b);
+                    Navigator.of(context).pop();
+                  },
+                ),
+              ],
+            ));
   }
 
   ListTile _buildItem(BuildContext context, int index) {
     Produto b = _lista[index];
 
     return ListTile(
-      leading: const Icon(Icons.propane), //TODO trocar esse icon pelo de produtos kkkk
+      leading: const Icon(
+          Icons.propane), //TODO trocar esse icon pelo de produtos kkkk
       title: Text('${b.id} '),
       subtitle: Text(b.descricao),
       onTap: () {
@@ -146,18 +160,19 @@ class _ListarProdutosPageState extends State<ListarProdutosPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: const Text("Listagem de Produtos"),
-        ),
-        drawer: const AppDrawer(),
-        body: ListView.builder(
-          itemCount: _lista.length,
-          itemBuilder: _buildItem,
-        ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () => Navigator.pushReplacementNamed(context, Routes.produtoInsert),
-          child: const Icon(Icons.add),
-        ),
+      appBar: AppBar(
+        title: const Text("Listagem de Produtos"),
+      ),
+      drawer: const AppDrawer(),
+      body: ListView.builder(
+        itemCount: _lista.length,
+        itemBuilder: _buildItem,
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () =>
+            Navigator.pushReplacementNamed(context, Routes.produtoInsert),
+        child: const Icon(Icons.add),
+      ),
     );
   }
 }
